See merge request ISPWA/intershop-pwa!641

+++ src/app/extensions/quoting/shared/quote/components/quote-edit/quote-edit.component.ts
@@ -1,5 +1,6 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, OnChanges, Output } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { Router } from '@angular/router';
 
 import { LineItemQuantity } from 'ish-core/models/line-item-quantity/line-item-quantity.model';
 import { User } from 'ish-core/models/user/user.model';
@@ -40,6 +41,7 @@ export class QuoteEditComponent implements OnChanges {
   @Output() submitQuoteRequest = new EventEmitter<void>();
   @Output() updateItem = new EventEmitter<LineItemQuantity>();
   @Output() deleteItem = new EventEmitter<string>();
+  @Output() deleteQuoteRequest = new EventEmitter<string>();
   @Output() copyQuote = new EventEmitter<void>();
   @Output() rejectQuote = new EventEmitter<void>();
   @Output() addQuoteToBasket = new EventEmitter<string>();
@@ -51,7 +53,7 @@ export class QuoteEditComponent implements OnChanges {
   validToDate: number;
   submitted = false;
 
-  constructor() {
+  constructor(private router: Router) {
     this.form = new FormGroup({
       displayName: new FormControl(undefined, [Validators.required, Validators.maxLength(255)]),
       description: new FormControl(undefined, []),
@@ -79,17 +81,29 @@ export class QuoteEditComponent implements OnChanges {
 
   /**
    * Throws updateItem event when onUpdateItem event trigggerd.
+   * Throws deleteQuoteRequest event when last items quantity is changed to '0'.
    * @param item Item id and quantity pair that should be changed
    */
   onUpdateItem(item: LineItemQuantity) {
-    this.updateItem.emit(item);
+    if (this.quote.items.length === 1 && item.quantity === 0) {
+      this.deleteQuoteRequest.emit(this.quote.id);
+      this.router.navigate(['/account/quote-list']);
+    } else {
+      this.updateItem.emit(item);
+    }
   }
 
   /**
    * Throws deleteItem event when delete button was clicked.
+   * Throws deleteQuoteRequest event when last item will be deleted.
    */
-  onDeleteItem(itemId) {
-    this.deleteItem.emit(itemId);
+  onDeleteItem(itemId: string) {
+    if (this.quote.items.length === 1) {
+      this.deleteQuoteRequest.emit(this.quote.id);
+      this.router.navigate(['/account/quote-list']);
+    } else {
+      this.deleteItem.emit(itemId);
+    }
   }
 
   /**
